Refactoring dataloaders
