+++ configs/base/fmow.yaml
@@ -9,27 +9,27 @@ checkpoint:
   filename: "{epoch:02d}-{val_acc_avg:.4f}-{val_acc_worst_region:.4f}-{val_loss:.4f}"
 
 dataset: 
-  class_reference: utils.data_utils.FMoWDataset
+  class_reference: data.fmow.FMoWDataset
   params:
     randaug: false
     clip_processor: openai/clip-vit-large-patch14-336
     group_field: "region"
-    # combine_id_ood: True
-    # split_groups:
-    #   train:
-    #     # - "Africa"
-    #     - "Americas"
-    #     - "Asia"
-    #     - "Europe"
-    #     - "Oceania"
-    #     - "Other"
-    #   val:
-    #     # - "Africa"
-    #     - "Americas"
-    #     - "Asia"
-    #     - "Europe"
-    #     - "Oceania"
-    #     - "Other"
+    combine_id_ood: False
+    split_groups:
+      train:
+        - "Africa"
+        - "Americas"
+        - "Asia"
+        - "Europe"
+        - "Oceania"
+        - "Other"
+      val:
+        - "Africa"
+        - "Americas"
+        - "Asia"
+        - "Europe"
+        - "Oceania"
+        - "Other"
 
 data_loader:
   batch_size: 16
